validate middleware

+++ utils/middleware/public_middleware.go
@@ -68,7 +68,7 @@ func PublicMiddleWare(extractionKey string, restKey string, publicAuthKey string
 		}
 
 		c.Set("middleware_type", Public)
-		c.Set("user_id", 0)
+		c.Set("user_id", int64(0))
 		c.Set("device_id", "")
 		c.Set("identity_key", "")
 		c.Next()
